requirements.txt.txt
# ===== Flask ve Web Framework =====
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
flask-restx==1.1.0
marshmallow==3.20.1
gunicorn==21.2.0
waitress==2.1.2

# ===== Machine Learning & Data Science =====
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2
scipy==1.11.1

# ===== Deep Learning & Computer Vision =====
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.165
opencv-python==4.8.0.76
Pillow==10.0.0
albumentations==1.3.1
imageio==2.31.1

# ===== YOLOv5 Dependencies =====
PyYAML>=5.3.1
matplotlib==3.7.2
seaborn==0.12.2
thop>=0.1.1

# ===== API ve HTTP =====
requests==2.31.0
python-dotenv==1.0.0
python-decouple==3.8

# ===== Veritabanı =====
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5

# ===== Logging & Monitoring =====
python-json-logger==2.0.7

# ===== Geolocation & Map =====
geopy==2.4.0
folium==0.14.0

# ===== Utilities =====
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
cryptography==41.0.3

# ===== Testing =====
pytest==7.4.0
pytest-flask==1.2.0

psycopg2